packages/agents/pyproject.toml
[project]
name = "chimera-agents"
version = "0.1.0"
description = "Python agents for AWS Chimera multi-tenant agent orchestration"
requires-python = ">=3.11"
dependencies = [
    "strands-agents>=1.0.0,<2.0.0",
    "bedrock-agentcore>=1.0.0,<2.0.0",
    "boto3>=1.34.0,<2.0.0",
    "botocore>=1.34.0,<2.0.0",
    "aiohttp>=3.9.0,<4.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-asyncio>=0.23.0,<1.0.0",
    "pytest-mock>=3.12.0,<4.0.0",
    "moto>=5.0.0,<6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = [
    "chimera_agent.py",
    "chat.py",
    "tools/",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: tests that require live AWS credentials (excluded from CI unit runs)",
]

[dependency-groups]
dev = []
